v0.0.2

+++ Backend/Api/Host/Endpoints/Responses/ProductResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace ModernTenon.Api.Host;
 
-public record ProductResponse(Guid Id, string Name, double? Price);
+public record ProductResponse(Guid Id, string Name, decimal? Price);
